Cleaning up _BasePCov along with new PCovR and PCovC subclasses
