+++ tests/test_kernel_pcovr.py
@@ -7,7 +7,11 @@ from sklearn.kernel_ridge import KernelRidge
 from sklearn.linear_model import Ridge, RidgeCV
 from sklearn.utils.validation import check_X_y
 
-from skmatter.decomposition import KernelPCovR, PCovR
+import sys
+sys.path.append('scikit-matter')
+from src.skmatter.decomposition.pcovr_new import PCovR
+from src.skmatter.decomposition._kernel_pcovr import KernelPCovR
+
 from skmatter.preprocessing import StandardFlexibleScaler as SFS
 
 
@@ -59,7 +63,6 @@ class KernelPCovRErrorTest(KernelPCovRBaseTest):
         for mixing in np.linspace(0, 1, 6):
             kpcovr = KernelPCovR(mixing=mixing, n_components=2, tol=1e-12)
             kpcovr.fit(self.X, self.Y)
-
             error = (
                 np.linalg.norm(self.Y - kpcovr.predict(self.X)) ** 2.0
                 / np.linalg.norm(self.Y) ** 2.0
