+++ tests/test_pcovr.py
@@ -10,8 +10,6 @@ from sklearn.linear_model import Ridge
 from sklearn.preprocessing import StandardScaler
 from sklearn.utils.validation import check_X_y
 
-#from skmatter.decomposition import PCovR
-
 import sys
 sys.path.append('scikit-matter')
 from src.skmatter.decomposition.pcovr_new import PCovR
